Preliminary Data Cleansing
Here's what I have so far with cleaning the data! The RAWG file cleaned up somewhat well, but the Steam data set seems to have a lot of scattered NaNs to the extent that trying to use dropna() eliminates all columns. I also did not work on the API because it required me to order it from some sort of form, so I thought it would be best if we discussed beforehand if we really wanted to use it. We can also discuss which features we will keep during our next meeting (I dropped a few that I knew we wouldn't want to use for sure, but some I thought might be worth looking into).